requirements.txt
beautifulsoup4==4.11.1
black==22.3.0
bs4==0.0.1
camelot-py==0.10.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==36.0.2
et-xmlfile==1.1.0
ghostscript==0.7
idna==3.3
mypy-extensions==0.4.3
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
pathspec==0.9.0
pdfminer.six==20220506
platformdirs==2.5.2
pycparser==2.21
PyPDF2==1.27.12
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
tabulate==0.8.9
tomli==2.0.1
urllib3==1.26.9
